big refactoring for rendering doc comments
