+++ lib/templates/_callable.html
@@ -7,5 +7,5 @@
     </span>
 </dt>
 <dd>
-    {{#oneLiner}}{{ documentation }}{{/oneLiner}}
+    {{{ oneLineDoc }}}
 </dd>
