+++ lib/templates/_constant.html
@@ -4,5 +4,5 @@
     <span class="constant-value">{{{ constantValue }}}</span>
 </dt>
 <dd>
-    {{#oneLiner}}{{ documentation }}{{/oneLiner}}
+    {{{ oneLineDoc }}}
 </dd>
