+++ lib/templates/_documentation.html
@@ -1,9 +1,7 @@
 {{#hasDocumentation}}
 <section class="desc markdown">
 
-    {{#markdown}}
-    {{{ documentation }}}
-    {{/markdown}}
+    {{{ documentationAsHtml }}}
 
 </section>
 {{/hasDocumentation}}
