+++ lib/templates/_head.html
@@ -38,7 +38,7 @@
                 {{{ layoutTitle }}}
             </h1>
             <p>
-                {{#oneLiner}}{{{ documentation }}}{{/oneLiner}}
+                {{{ oneLineDoc }}}
             </p>
         </div>
     </div>
