+++ lib/templates/_property.html
@@ -5,5 +5,5 @@
 </dt>
 <dd>
     {{>readable_writable}}
-    {{#oneLiner}}{{ documentation }}{{/oneLiner}}
+    {{{ oneLineDoc }}}
 </dd>
