+++ lib/templates/class.html
@@ -75,7 +75,7 @@
                     const
                 </div>
                 {{/isConst}}
-                {{#oneLiner}}{{ documentation }}{{/oneLiner}}
+                {{{ oneLineDoc }}}
             </dd>
             {{/class.constructors}}
         </dl>
