+++ lib/templates/index.html
@@ -12,7 +12,7 @@
               <span class="name"><a href="{{nameForFile}}/index.html">{{ name }}</a></span>
             </dt>
             <dd>
-              {{#oneLiner}}{{{ documentation }}}{{/oneLiner}}
+                {{{ oneLineDoc }}}
             </dd>
           {{/package.libraries}}
         </dl>
