+++ lib/templates/library.html
@@ -15,7 +15,7 @@
                 <span class="name {{#isDeprecated}}deprecated{{/isDeprecated}}">{{{linkedName}}}</span>
             </dt>
             <dd>
-                {{#oneLiner}}{{{ documentation }}}{{/oneLiner}}
+                {{{ oneLineDoc }}}
             </dd>
             {{/library.classes}}
         </dl>
@@ -33,7 +33,7 @@
                 {{{linkedName}}}
             </dt>
             <dd>
-                {{#oneLiner}}{{ documentation }}{{/oneLiner}}
+                {{{ oneLineDoc }}}
             </dd>
             {{/library.enums}}
         </dl>
@@ -98,7 +98,7 @@
                 <span class="name {{#isDeprecated}}deprecated{{/isDeprecated}}">{{{linkedName}}}</span>
             </dt>
             <dd>
-                {{#oneLiner}}{{ documentation }}{{/oneLiner}}
+                {{{ oneLineDoc }}}
             </dd>
             {{/library.exceptions}}
         </dl>
